remove the CT sensor based solution for detecting presence
First problem I found was that CT sensors have problems with noise due to manufacturing and design problem (https://www.casler.org/wordpress/low-current-measurement-performance-of-the-sct013/), then I found that the HAT for the CT sensors has a Current range as well and the last point was that I found out that the CT sensor cannot be used only for measuring single lead wire current. Thus, I needed to find a different solution, possibly based on smart sockets.